feat: it is possible to create a game using the UI
