+++ package.json
@@ -52,6 +52,7 @@
     "firebase": "^6.2.0",
     "lit-element": "^2.0.1",
     "lit-html": "^1.0.0",
+    "pwa-helpers": "^0.9.1",
     "redux": "^4.0.1",
     "uuid": "^3.3.2"
   }
